pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "beacon-io"
version = "0.1.0"
description = "BEACON-IO: Expression-driven dependencies at the tumor-immune interface for precision immuno-oncology"
requires-python = ">=3.10,<3.13"
license = {text = "MIT"}
authors = [
    {name = "Abdulkadir Elmas"},
]
dependencies = [
    # Core data
    "pandas>=2.0",
    "numpy>=1.24",
    "scipy>=1.11",
    "anndata>=0.10",
    "scanpy>=1.10",
    # Bayesian inference (BEACON engine)
    "pymc>=5.10",
    "arviz>=0.17",
    # Statistics / ML
    "statsmodels>=0.14",
    "scikit-learn>=1.3",
    "lifelines>=0.28",          # survival analysis
    # "scikit-survival>=0.22",  # penalised Cox models (optional, needs compiler)
    # Immune deconvolution
    "decoupler>=1.6",           # TF/pathway activity inference
    # Networks
    "networkx>=3.1",
    # Genomics / bioinformatics
    "pydeseq2>=0.4",            # differential expression
    # Data access
    "synapseclient>=4.0",       # Synapse (for ICB cohorts)
    "requests>=2.28",
    # Visualisation
    "matplotlib>=3.7",
    "seaborn>=0.13",
    "plotnine>=0.13",
    "upsetplot>=0.9",
    # Utilities
    "pyyaml>=6.0",
    "tqdm>=4.65",
    "rich>=13.0",
]

[project.optional-dependencies]
singlecell = [
    "scvi-tools>=1.0",         # scVI latent-space integration
]
gpu = [
    "jax[cuda12]>=0.4",        # GPU acceleration for PyMC
]
snakemake = ["snakemake>=8.0"]
dev = ["pytest", "ruff", "ipykernel", "jupyter", "pre-commit"]
all = [
    "beacon-io[singlecell,snakemake,dev]",
]

# CLI entry point (uncomment when cli.py is created)
# [project.scripts]
# beacon-io = "beacon_io.cli:main"

[tool.setuptools.packages.find]
where = ["src"]
include = ["beacon_io*"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "W", "UP"]

[tool.pytest.ini_options]
testpaths = ["tests"]
